otto-cases-backend/requirements.txt
fastapi
uvicorn
supabase
openai
jinja2
weasyprint
pydantic
python-dotenv
firebase-admin
google-cloud-firestore
google-cloud-storage
python-multipart
